airflow/requirements.txt
kafka-python-ng
apache-airflow
psycopg2-binary
